refactor: rename PDF generation function and update invoice validation logic to support multiple invoice types
