+++ ied-be/src/routes/racuni_v2.routes.ts
@@ -1,6 +1,6 @@
 import { type Request, type Response, Router } from "express";
 import { type RacunV2Form, RacunV2Zod, TipRacuna } from "ied-shared";
-import { generatePredracunV2Pdf } from "../services/racuni_v2.service";
+import { generateRacunV2Pdf } from "../services/racuni_v2.service";
 import { getTemplateErrorDetails } from "../utils/docx.utils";
 
 const router = Router();
@@ -19,9 +19,13 @@ router.post(
         return;
       }
 
-      if (parsed.data.tipRacuna !== TipRacuna.PREDRACUN) {
+      if (
+        parsed.data.tipRacuna !== TipRacuna.PREDRACUN &&
+        parsed.data.tipRacuna !== TipRacuna.AVANSNI_RACUN
+      ) {
         res.status(400).json({
-          message: "Racun V2 PDF POC currently supports only Predracun.",
+          message:
+            "Racun V2 PDF POC currently supports only Predracun and Avansni racun.",
           tipRacuna: parsed.data.tipRacuna,
         });
         return;
@@ -35,7 +39,7 @@ router.post(
         return;
       }
 
-      const { buffer, fileName } = await generatePredracunV2Pdf(parsed.data);
+      const { buffer, fileName } = await generateRacunV2Pdf(parsed.data);
       res.setHeader("Content-Type", "application/pdf");
       res.setHeader("Content-Disposition", `attachment; filename=${fileName}`);
       res.send(buffer);
