+++ ied-be/src/services/racuni_v2.service.ts
@@ -2,30 +2,43 @@ import {
   calcInvoiceTotals,
   isIzdavacPdvObveznik,
   type RacunV2Parsed,
+  TipRacuna,
 } from "ied-shared";
 import { izdavacRacuna } from "../constants/izdavacRacuna.const";
 import {
+  type InvoiceTemplateData,
+  type InvoiceTemplateName,
   type IzdavacProfile,
-  type PredracunTemplateData,
   type PrimalacView,
-  renderPredracun,
-} from "../templates/predracun.template";
+  renderInvoice,
+} from "../templates/invoiceRenderer";
 import { sanitizeFilename } from "../utils/docx.utils";
 import { env } from "../utils/envVariables";
 import { renderHtmlToPdfBuffer } from "./docx.service";
 
-type PredracunV2Parsed = Extract<
-  RacunV2Parsed,
-  { tipRacuna: "predracun" }
->;
-
 type IzdavacConfig = (typeof izdavacRacuna)[number];
 
-export const generatePredracunV2Pdf = async (
-  racunData: PredracunV2Parsed,
+/**
+ * Maps each supported `tipRacuna` to the Handlebars template it renders.
+ * Add an entry when a new invoice variant ships (konacni, racun); the rest
+ * of the pipeline (view-model builder, file-name, PDF render) is shared.
+ *
+ * `Partial<Record<…>>` because unsupported variants throw upstream — both
+ * the route's allow-list and {@link pickTemplateName}'s explicit check.
+ */
+const TEMPLATE_NAME_BY_TIP_RACUNA: Partial<
+  Record<TipRacuna, InvoiceTemplateName>
+> = {
+  [TipRacuna.PREDRACUN]: "predracun",
+  [TipRacuna.AVANSNI_RACUN]: "avansni",
+};
+
+export const generateRacunV2Pdf = async (
+  racunData: RacunV2Parsed,
 ): Promise<{ buffer: Buffer; fileName: string }> => {
-  const viewModel = buildPredracunTemplateData(racunData);
-  const html = renderPredracun(viewModel, { css: "inline" });
+  const templateName = pickTemplateName(racunData.tipRacuna);
+  const viewModel = buildInvoiceTemplateData(racunData);
+  const html = renderInvoice(templateName, viewModel, { css: "inline" });
   const buffer = await renderHtmlToPdfBuffer(html, {
     format: "A4",
     margin: { top: 0, right: 0, bottom: 0, left: 0 },
@@ -33,13 +46,34 @@ export const generatePredracunV2Pdf = async (
 
   return {
     buffer,
-    fileName: getPredracunFileName(racunData),
+    fileName: getInvoiceFileName(racunData),
   };
 };
 
-const buildPredracunTemplateData = (
-  racunData: PredracunV2Parsed,
-): PredracunTemplateData => {
+const pickTemplateName = (tipRacuna: TipRacuna): InvoiceTemplateName => {
+  const name = TEMPLATE_NAME_BY_TIP_RACUNA[tipRacuna];
+  if (!name) {
+    // Defense-in-depth: the route's allow-list already rejects unsupported
+    // variants with a 400. This throw turns a programmer mistake (forgetting
+    // to update the map when a new tip lands) into a 500 with a clear
+    // message instead of a confusing Handlebars "file not found".
+    throw new Error(
+      `Racun V2 PDF generation not supported for tipRacuna: ${tipRacuna}`,
+    );
+  }
+  return name;
+};
+
+/**
+ * Shared view-model builder for every invoice variant. Most fields are
+ * identical across `RacunV2Parsed` branches; the only branch-specific bits
+ * are the two mutually exclusive payment-date fields (`rokZaUplatu` on
+ * predracun/konacni/racun, `datumUplateAvansa` on avansni). The template
+ * ignores whichever is absent.
+ */
+const buildInvoiceTemplateData = (
+  racunData: RacunV2Parsed,
+): InvoiceTemplateData => {
   const pdvObveznik = isIzdavacPdvObveznik(racunData.izdavacRacuna);
   const totals = calcInvoiceTotals(
     racunData.stavke,
@@ -51,7 +85,12 @@ const buildPredracunTemplateData = (
     pozivNaBroj: racunData.pozivNaBroj,
     datum: new Date(),
     mesto: "Beograd",
-    rokZaUplatu: racunData.rokZaUplatu,
+    rokZaUplatu:
+      "rokZaUplatu" in racunData ? racunData.rokZaUplatu : undefined,
+    datumUplateAvansa:
+      racunData.tipRacuna === TipRacuna.AVANSNI_RACUN
+        ? racunData.datumUplateAvansa
+        : undefined,
     valuta: racunData.valuta,
     pdvObveznik,
     izdavac: buildIzdavacProfile(racunData),
@@ -63,9 +102,7 @@ const buildPredracunTemplateData = (
   };
 };
 
-const buildIzdavacProfile = (
-  racunData: PredracunV2Parsed,
-): IzdavacProfile => {
+const buildIzdavacProfile = (racunData: RacunV2Parsed): IzdavacProfile => {
   const issuer = izdavacRacuna.find(
     (item) => item.id === racunData.izdavacRacuna,
   );
@@ -94,7 +131,7 @@ const splitIzdavacNaziv = (
   };
 };
 
-const buildPrimalacView = (racunData: PredracunV2Parsed): PrimalacView => {
+const buildPrimalacView = (racunData: RacunV2Parsed): PrimalacView => {
   const primalac = racunData.primalacRacuna;
   const adresa = joinAddress(primalac.adresa, primalac.mesto);
 
@@ -125,7 +162,7 @@ const joinAddress = (...parts: Array<string | undefined>): string => {
 
 const parseTekuciRacun = (
   tekuciRacun: string,
-): PredracunTemplateData["uplata"] => {
+): InvoiceTemplateData["uplata"] => {
   const match = tekuciRacun.match(/^(.*?)\s*\((.*?)\)\s*$/);
   if (!match) {
     return {
@@ -148,7 +185,7 @@ const parseTekuciRacun = (
   };
 };
 
-const getPredracunFileName = (racunData: PredracunV2Parsed): string => {
+const getInvoiceFileName = (racunData: RacunV2Parsed): string => {
   const primalac = racunData.primalacRacuna;
   const primalacNaziv =
     primalac.tipPrimaoca === "firma" ? primalac.naziv : primalac.imeIPrezime;
