+++ ied-be/src/templates/invoices/avansni.hbs
@@ -29,12 +29,14 @@
             <tr>
               <th scope="col">Datum</th>
               <th scope="col">Mesto</th>
+              <th scope="col">Uplata avansa</th>
             </tr>
           </thead>
           <tbody>
             <tr>
               <td>{{formatDate datum}}</td>
               <td>{{mesto}}</td>
+              <td>{{formatDate datumUplateAvansa}}</td>
             </tr>
           </tbody>
         </table>
